Fix a bug in timezone handling in dt2age.
Mongs has a function to convert datetime objects or timestamps to an age
representation like "2 hours ago." Its first step is to compute the number of
seconds between the current time and the time given. When passing in a datetime
object it cast this to a timestamp in the local time, but the subsequent
subtraction is from the current timestamp in UTC, resulting in erroneous
representations of age in the amount of the different between local time and
UTC. In fixing this bug, the opportunity was taken to remove a dependency on
the time module in favor of the more modern datetime module.