Work around collection_names for many collections
The collection_names API can block your server for minutes or hours if
you have a great number of collections. I ran it against a server with
~85,000 collections, and it blocked the server for 15 minutes before it
was restarted. On the other hand, MongoDB stores collection information
in a meta-collection called system.namespaces. I've special-cased this
collection to be minimally browseable with Mongs, linking to individual
collections, and I'm redirecting the collections listing to
system.namespaces if there are more than 256 namespace items (which
include indexes, which are themselves special collections).